# Optimize raw_next()