Even bench_new is faster because nothing is allocated on the heap right away.

**Before**
```
test tests::bench_are_all_chars_last ... bench:           8 ns/iter (+/- 0) (no longer used)
test tests::bench_new                ... bench:          74 ns/iter (+/- 4)
test tests::bench_next               ... bench:         303 ns/iter (+/- 6)
test tests::bench_next_char          ... bench:           1 ns/iter (+/- 0) (no longer used)
test tests::bench_raw_next           ... bench:         238 ns/iter (+/- 20)
```

**After**
```
test tests::bench_new      ... bench:           9 ns/iter (+/- 0)
test tests::bench_next     ... bench:          84 ns/iter (+/- 5)
test tests::bench_raw_next ... bench:          27 ns/iter (+/- 2)
```